scripts/requirements.txt
requests>=2.32.5
playwright>=1.58.0
PyYAML>=6.0
